Code of Stacked AE
Stacked AutoEncoder is implemented on KDD99 dataset. The full data "kddcup.data.corrected" is used to TRAIN the AE and "kddcup.data_10_percent_corrected" is used to TEST the AE. The "test_loss"  is the final margin between observed and actual values. The NN is trained on the basis of Prototype tool, service and result.